requirements.txt
matplotlib==3.10.8
numpy==2.4.4
opencv-python==4.13.0.84
Pillow==12.2.0
